meta: Rename to rust-container
I originally wanted to rewrite nixos-container in Rust as a way to
practice writing Rust code (and understand more about nixos-container),
however now that I *have* delved into the source code, it's apparent
that a lot of the code in nixos-container is legacy code and not
representative of how people usually manage containers in 2023.

For this reason, this project will not be a rewrite, but rather a
solution for how I want to use containers specifically.