meta: Remove sakaya-gui
Although this was a fun learning exercise, maintaining a GUI for sakaya
seems out of scope when there are alternative solutions that leverage
existing GUI implementations.

For example, it's possible to skip the GUI altogether and only launch
applications directly from the user's file manager of choice. A user may
also have an application launcher that uses the desktop files created by
sakaya, which seems preferable over the user having to navigate a
separate window that may or may not conform to their ideals.